Move embed components to components folder
