+++ src/index.js
@@ -1,1 +1,2 @@
 import './blocks/activity';
+import './blocks/route';
